+++ ProjetStockManager/index.html
@@ -10,7 +10,6 @@
     <script type="module" src="./js/fonctionsDeGestion.js" defer></script>
     <script type="module" src="./js/messageutilisateur.js" defer></script>
     <script type="module" src="./js/scriptPrincipal.js" defer></script>
-    <script type="module" src="./js/test.js" defer></script>
     <script type="module" src="./js/index.js" defer></script>
     <link rel="stylesheet" href="./css/styleGénéral.css"/>
     <title>Document</title>
