+++ ProjetStockManager/js/fonctionsDeBDD.js
@@ -1,20 +1,22 @@
 import { ajouterUnObjetAvecIdSpécifique  } from "./fonctionsCRUDFirebase.js";
-import { tableauObjectDeChamp } from "./test.js";
+import { tableauObjectDeChamp } from "./variablesGlobales.js";
 import { objetConstructeur } from "./fonctionsContruction.js"
 
 // Function pour le bouton "croix-valider" qui declenche les fonctions base sur les inputs.
-let ajouterCollection = document.getElementById("ajout_BDD");
-ajouterCollection.addEventListener("click", () => {
-  let obj = creerCollection(tableauObjectDeChamp);
-  console.log(obj)
-  objetConstructeur(tableauObjectDeChamp); 
-});
+
+const ajouterUneCollection = () => {
+  let ajouterCollection = document.getElementById("ajout_BDD");
+  ajouterCollection.addEventListener("click", () => {
+    objetConstructeur(tableauObjectDeChamp); 
+  });
+}
+
 
 // Creer un objet en prenant les valeurs des champs de text
 const creerCollection = (tableauObjectDeChamp) => {
   let objet = {};
   for (let obj of tableauObjectDeChamp) {
-    if(document.getElementById(obj.nom)){
+    if(document.getElementById(obj.nom)) {
       objet[obj.nom] = document.getElementById(obj.nom).value;
     }
   }
@@ -38,6 +40,6 @@ const creerCollection = (tableauObjectDeChamp) => {
 
 
 export {
+  ajouterUneCollection,
   creerCollection,
-  ajouterCollection,
 }
