+++ ProjetStockManager/js/variablesGlobales.js
@@ -2,7 +2,7 @@ let tableauObjectDeChamp = [
     {nom:"Nom", type: "text"},
     {nom:"Référence", type: "text"},
     {nom:"Quantité", type: "number", step:"1"},
-    {nom:"Catégorie", type: "text", check: false},
+    {nom:"Catégorie", type: "text"},
     {nom:"Prix", type: "number", step:"0.01", check:false},
     {nom:"Date", type: "date", check:false},
     {nom:"SousCatégorie", type: "text", check:false},
